requirements/base.txt
alabaster==0.7.10
Babel==2.5.1
beautifulsoup4==4.6.0
boto3==1.4.7
botocore==1.7.12
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
Collectfast==0.2.3
coverage==4.2
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.5
django-allauth==0.26.1
django-anymail==0.4.2
django-autoslug==1.9.3
django-braces==1.9.0
django-coverage-plugin==1.3.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.5
django-environ==0.4.0
django-extensions==1.7.1
django-floppyforms==1.6.2
django-model-utils==2.5.2
django-modelcluster==3.1
django-redis==4.4.4
django-storages==1.6.5
django-storages-redux==1.3.2
django-taggit==0.22.1
django-test-plus==1.0.14
django-treebeard==4.1.2
djangorestframework==3.6.4
docutils==0.14
factory-boy==2.7.0
fake-factory==9999.9.9
gevent==1.1.2
greenlet==0.4.12
gunicorn==19.6.0
houdini.py==0.1.0
html5lib==0.999999999
idna==2.6
imagesize==0.7.1
ipdb==0.10.1
ipython==6.2.0
ipython-genutils==0.2.0
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.3
libsass==0.13.2
MarkupSafe==1.0
misaka==2.1.0
oauthlib==2.0.4
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
psycopg2==2.6.2
ptyprocess==0.5.2
py==1.4.34
pycparser==2.18
Pygments==2.2.0
pytest==3.2.2
pytest-django==2.9.1
pytest-sugar==0.7.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.11
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.4.5
sqlparse==0.2.3
termcolor==1.1.0
traitlets==4.3.2
unicode-slugify==0.1.3
Unidecode==0.4.21
urllib3==1.22
wagtail==1.12.1
Wand==0.4.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.11.10
Willow==0.4
